ep003.h: int now long int

+++ C++/problem003/ep003.h
@@ -2,12 +2,12 @@
 #define _ep003_h
 
 // function which returns the largest prime factor of a target, x 
-int LargestPrimeFactor(int x);
+long int LargestPrimeFactor(long int x);
 
-// test if a number is integer
-bool IsFactor(int target, int factor);
+// test if factor is a factor of target
+bool IsFactor(long int target, long int factor);
 
 // test if number is prime
-bool IsPrime(int x);
+bool IsPrime(long int x);
 
 #endif
